requirements.txt
kafka-python
psycopg2-binary
prometheus_client